Updates for version 1.1 beta 1.
This update include the following new features.

Multiplayer - This API provides support for setting up games and lobbies between players on different devices. Plus, you can use the API to send commands to a remote Sphero connected to another device. It works over WiFi and it can communicate with applications built with the Sphero SDK for other platforms.

Collision Detection - This API provides support for detecting collisions as registered by the Sphero firmware. Version 1.13 or greater firmware is needed to use this feature.

Macros - Provides support to run macros on Sphero. Macros are a series of command in binary form that can be downloaded and ran on Sphero. The API supports the dynamic building of macros. See the Orbotix MacroLab application to test out programing a macro and playing them. Macros built with MacroLab can be saved and used in your application.

Currently the API documentation has not been updated, but the frameworks contains headers, which are commented.